Stricter warnings (#86)
* stricter warnings
* fix warning for const return type
* ignore warnings related to anonymous types
* turn on warnings as errors for gcc/clang
* fix warning for unused variable
* ignore nameless struct/union extension for MSVC
* fix warnings in lib
* fix warnings in tests
* fix gcc error
* set up warnings-as-error to be configurable
* add warning to test that GHA will succeed even with warning
* set warnings as errors in GHA build
* remove intentional warning to see if GHA passes
* suppress warnings in code instead of cmake
* granular warning suppression
* move cmake options deeper into the build tree

+++ code/src/stf/include/interface/stf/geom/aabb.hpp
@@ -93,7 +93,7 @@ namespace stf::geom
             vec_t vert = min;
             for (size_t j = 0; j < N; ++j)
             {
-                if ((i & (1 << j)) != 0)
+                if ((i & (static_cast<size_t>(1) << j)) != 0)
                 {
                     vert[j] = max[j];
                 }
